[2 - 4 단계 방탈출 예약 결제 / 배포] 비토(오상훈) 미션 제출합니다. (#100)
* refactor: 2단계 클라이언트 코드 반영

* feat: 나의 예약 페이지에 paymentKey 및  가격 출력 기능 구현

* refactor: 출력 형식 변경

* refactor: Entity 연관관계 수정

* feat: swagger 적용

* feat: 예약에 대한 논리적 삭제 적용

* refactor: 결제 정보 선 저장 후 결제하도록 수정

* fix: 결제 정보가 있는 예약 삭제 오류 수정

* refactor: 메서드 명 수정

* refactor: AbstractExceptionHandler 삭제

* refactor: 불필요한 메서드(getPrintStatus) 삭제

* refactor: Transactional 어노테이션 적용

* refactor: 예약 삭제 시 환불 기능 구현

* refactor: 커스텀 예외의 매개변수에 HttpStatus 삭제

* refactor: admin 페이지에 페이지 이동 버튼 추가

* refactor: 예약 삭제 시 자신의 예약인지 확인하는 기능 구현

* refactor: 클래스 이름 변경

* refactor: 상수명 및 메서드 명 변경

* fix: 오류 수정

* refactor: 예약 취소 상태로 추가

* fix: 예약 삭제 시 첫번째 예약 대기 자동 승인 기능 오류 수정

+++ src/main/java/roomescape/dto/request/reservation/ReservationInformRequest.java
@@ -1,5 +1,6 @@
 package roomescape.dto.request.reservation;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ReservationInformRequest(
@@ -9,6 +10,6 @@ public record ReservationInformRequest(
         long themeId,
         String paymentKey,
         String orderId,
-        int amount
+        BigDecimal amount
 ) {
 }
